requirements.txt
opencv-python
numpy
gTTS
moviepy
Pillow
requests
